requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0

# Machine learning
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
imbalanced-learn>=0.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Data acquisition
requests>=2.31.0

# Model persistence
joblib>=1.3.0

# Progress bars (optional but nice)
tqdm>=4.65.0
